pyproject.toml
# Author: Goutam Adwant
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "kvfleet"
version = "0.9.0"
description = "Production-grade, KV-cache-aware intelligent routing for self-hosted and hybrid LLM fleets."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Goutam Adwant" }]
keywords = ["llm-routing", "kv-cache", "model-fleet", "self-hosted-llm", "gpu-routing", "inference-routing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "typer>=0.16.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.2.0",
    "twine>=6.0.0",
    "pre-commit>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "hypothesis>=6.119.0",
    "pip-audit>=2.8.0",
    "respx>=0.22.0",
]
semantic = [
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]
gateway = [
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
]

[project.urls]
Homepage = "https://github.com/adwantg/kvfleet"
Repository = "https://github.com/adwantg/kvfleet"
Issues = "https://github.com/adwantg/kvfleet/issues"

[project.scripts]
kvfleet = "kvfleet.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kvfleet"]
core-metadata-version = "2.3"

[tool.hatch.build.targets.sdist]
core-metadata-version = "2.3"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers --strict-config --cov=kvfleet --cov-report=term-missing"
xfail_strict = true
asyncio_mode = "auto"

[tool.coverage.run]
source = ["kvfleet"]

[tool.coverage.report]
fail_under = 90
show_missing = true
skip_covered = true
